functions in the plugins are all updated and negui_general_plugin.py is now deprecated
