+++ test-app/src/plugins/negui_data_export_plugin.py
@@ -1,7 +1,7 @@
 import json
 import sbmlplot
 
-def dataExportInfo():
+def items_info():
     # json
     json_format = {'name' : "as JSON", 'type': "dataexporttool", 'file-extension' : "json", 'default-save-file-name' : "network", 'element-types': {'node': ["Reaction", "Compartment"], 'edge': ["Reactant", "Product", "Modifier"]}}
 
@@ -11,28 +11,29 @@ def dataExportInfo():
     return json.dumps({'items': [json_format, sbml_format]})
 
 
-def checkForGraphInfoCompatibility(graphInfoString, filetype):
+def checkForGraphInfoCompatibility(input):
+    graph_info = json.loads(input[0])
+    file_type = input[1]
     # json
-    if filetype == "as JSON":
-        return checkForJSONCompatibility(graphInfoString)
+    if file_type == "as JSON":
+        return checkForJSONCompatibility()
     # sbml
-    elif filetype == "as SBML":
-        return checkForSBMLCompatibility(graphInfoString)
+    elif file_type == "as SBML":
+        return checkForSBMLCompatibility(graph_info)
 
-    return json.dumps({'isInfoCompatible': False, 'messages': [{'message': "export format is not supported"}]})
+    return ({'isInfoCompatible': False, 'messages': [{'message': "export format is not supported"}]},)
 
-def checkForJSONCompatibility(graphInfoString):
-    return json.dumps({'isInfoCompatible': True})
 
-def checkForSBMLCompatibility(graphInfoString):
+def checkForJSONCompatibility():
+    return (json.dumps({'isInfoCompatible': True}),)
+
+
+def checkForSBMLCompatibility(graph_info):
     all_species_have_parents = True
-    # read graph info
-    graphInfo = json.loads(graphInfoString)
-
-    for n_index in range(len(graphInfo['nodes'])):
-        if graphInfo['nodes'][n_index]['style']['category'] == "Species":
-            if 'parent' not in list(graphInfo['nodes'][n_index].keys()) \
-                    or graphInfo['nodes'][n_index]['parent'] == "N/A":
+    for n_index in range(len(graph_info['nodes'])):
+        if graph_info['nodes'][n_index]['style']['category'] == "Species":
+            if 'parent' not in list(graph_info['nodes'][n_index].keys()) \
+                    or graph_info['nodes'][n_index]['parent'] == "N/A":
                 all_species_have_parents = False
 
     messages = []
@@ -40,29 +41,32 @@ def checkForSBMLCompatibility(graphInfoString):
         messages.append({'message': "Some Species in the network do not belong to any Compartments."})
 
     is_info_compatible = all_species_have_parents
-    return json.dumps({'isInfoCompatible': is_info_compatible, 'messages': messages})
+    return (json.dumps({'isInfoCompatible': is_info_compatible, 'messages': messages}),)
 
-def writeGraphInfoToFile(graphInfoString, filename, filetype):
-    graphInfo = json.loads(graphInfoString)
+def writeGraphInfoToFile(input):
+    graph_info = json.loads(input[0])
+    file_type = input[1]
+    file_name = input[2]
 
     # json
-    if filetype == "as JSON":
-        writeJSON(graphInfo, filename)
+    if file_type == "as JSON":
+        return writeJSON(graph_info, file_name)
     # sbml
-    elif filetype == "as SBML":
-        writeSBML(graphInfo, filename)
+    elif file_type == "as SBML":
+        return writeSBML(graph_info, file_name)
 
 
-def writeJSON(graphInfo, filename):
-    with open(filename, 'w', encoding='utf8') as js_file:
-        json.dump(graphInfo, js_file, indent=1)
+def writeJSON(graph_info, file_name):
+    with open(file_name, 'w', encoding='utf8') as js_file:
+        json.dump(graph_info, js_file, indent=1)
 
-def writeSBML(graphInfo, filename):
+
+def writeSBML(graph_info, file_name):
     sbml_graph_info = sbmlplot.SBMLGraphInfoImportFromNetworkEditor()
-    sbml_graph_info.extract_info(graphInfo)
+    sbml_graph_info.extract_info(graph_info)
     sbml_export = sbmlplot.SBMLGraphInfoExportToSBMLModel()
     sbml_export.extract_graph_info(sbml_graph_info)
-    sbml_export.export(filename)
+    sbml_export.export(file_name)
 
 """
 # keep commented till using the new version of sbmlplot
